big refactor of workout class
