+++ rower_monitor/test.py
@@ -9,7 +9,7 @@ import matplotlib.pyplot as plt
 #csv_source = ds.CsvFile(None, "C:\\Users\\checo\\Desktop\\rower\\2020-06-22 21h39m58s.csv")
 #csv_source = ds.CsvFile(None, "C:\\Users\\checo\\Desktop\\rower\\2020-06-20 17h25m29s.csv")
 csv_source = ds.CsvFile(
-    "C:\\Users\\checo\\Desktop\\rower\\2020-06-30 09h20m55s.csv",
+    "C:\\Users\\checo\\Desktop\\rower\\2020-08-24 22h18m35s.csv",
     threaded=False,
     sample_delay=False,
 )
@@ -19,30 +19,28 @@ workout.start()
 workout.stop()
 
 
-print(workout.num_flywheel_revolutions)
+print(workout.boat.position.values[-1])
 
-work = [x.work_done_by_person for x in workout.strokes.values]
-plt.plot(workout.torque.timestamps, workout.torque.values)
-plt.scatter(workout.strokes.timestamps, work, s=1, c='red')
+work = [x.work_done_by_person / 5.0 for x in workout.person.strokes.values]
+plt.plot(workout.person.torque.timestamps, workout.person.torque.values)
+plt.scatter(workout.person.strokes.timestamps, work, s=1, c='red')
 plt.show()
 
 
 print('done')
-exit()
+blaa
 
 
-plt.plot(workout.acceleration.timestamps, workout.acceleration.values)
+plt.plot(workout.machine.flywheel_acceleration.timestamps, workout.machine.flywheel_acceleration.values)
 plt.show()
 
 
-plt.scatter(workout.speed.values[1:], workout.acceleration.values, s=1)
+plt.scatter(workout.machine.flywheel_speed.values[1:], workout.machine.flywheel_acceleration.values, s=1)
 plt.show()
 
-intercepts = [x.fitted_damping_model.intercept for x in workout.strokes.values]
-slopes = [x.fitted_damping_model.slope for x in workout.strokes.values]
-plt.plot(intercepts)
-#plt.plot(slopes)
+intercepts = [x.intercept for x in workout.machine.damping_models]
+slopes = [x.slope for x in workout.machine.damping_models]
+#plt.plot(intercepts)
+plt.plot(slopes)
 plt.show()
 
-plt.plot(workout.torque.timestamps, workout.torque.values)
-plt.show()
